refactor: move MockEntryPoint to independent file
* this change makes it possible to import the MockEntryPoints from an
  independend file. this class is used in two tests but defined in one
  of them. placing it into a separated file makes it independent.
